set background of html card webview to black on instantiation
The previous commit (92cfdafcfe180b5774e0935f3033f0f4d7899e3b) that
tried to take care of this missed the stage of loading where the view
has been instantiated but the body not loaded yet. When the body is
small this stage is of negligible length, but it shows up when the
body is big.